rename Go package "push" -> "deploy"
This allows us to have a sane command list, where "image_push", "image_load", and "image_deploy" all use "img deploy ..." (at some point).
