+++ img_tool/cmd/deploy/BUILD.bazel
@@ -1,9 +1,9 @@
 load("@rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "push",
-    srcs = ["push.go"],
-    importpath = "github.com/bazel-contrib/rules_img/img_tool/cmd/push",
+    name = "deploy",
+    srcs = ["deploy.go"],
+    importpath = "github.com/bazel-contrib/rules_img/img_tool/cmd/deploy",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/api",
