+++ img_tool/cmd/deploy/deploy.go
@@ -1,4 +1,4 @@
-package push
+package deploy
 
 import (
 	"bytes"
